+++ src/Exception/UnexpectedAttributeException.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Hydrator;
+namespace Yiisoft\Hydrator\Exception;
 
 use InvalidArgumentException;
 use Throwable;
