+++ src/Internal/ConstructorArgumentsExtractor.php
@@ -2,9 +2,15 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Hydrator;
+namespace Yiisoft\Hydrator\Internal;
 
 use ReflectionClass;
+use Yiisoft\Hydrator\Data;
+use Yiisoft\Hydrator\Hydrator;
+use Yiisoft\Hydrator\ParameterAttributesHandler;
+use Yiisoft\Hydrator\Result;
+use Yiisoft\Hydrator\TypeCastContext;
+use Yiisoft\Hydrator\TypeCasterInterface;
 
 /**
  * @internal
