+++ src/Internal/DataHelper.php
@@ -2,8 +2,9 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Hydrator;
+namespace Yiisoft\Hydrator\Internal;
 
+use Yiisoft\Hydrator\Result;
 use Yiisoft\Strings\StringHelper;
 
 use function is_array;
