+++ tests/Support/Attribute/FromPredefinedArrayResolver.php
@@ -11,7 +11,7 @@ use Yiisoft\Hydrator\DataAttributeResolverInterface;
 use Yiisoft\Hydrator\ParameterAttributeInterface;
 use Yiisoft\Hydrator\ParameterAttributeResolverInterface;
 use Yiisoft\Hydrator\Result;
-use Yiisoft\Hydrator\UnexpectedAttributeException;
+use Yiisoft\Hydrator\Exception\UnexpectedAttributeException;
 
 use function array_key_exists;
