+++ tests/UnexpectedAttributeExceptionTest.php
@@ -8,7 +8,7 @@ use LogicException;
 use PHPUnit\Framework\TestCase;
 use stdClass;
 use Yiisoft\Hydrator\Attribute\Parameter\DiResolver;
-use Yiisoft\Hydrator\UnexpectedAttributeException;
+use Yiisoft\Hydrator\Exception\UnexpectedAttributeException;
 
 final class UnexpectedAttributeExceptionTest extends TestCase
 {
